requirements.txt
flask
python-docx
pywin32
spacy